service20/requirements-api.txt
# FastAPI and ASGI Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
asyncpg>=0.29.0

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Environment Variables
python-dotenv>=1.0.1

# Observability - OpenTelemetry
opentelemetry-sdk>=1.20.0
opentelemetry-api>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0

# Observability - Arize Phoenix
arize-phoenix>=5.0.0
arize>=7.0.0
arize-otel>=0.1.0
openinference-instrumentation-langchain>=0.1.0

# Observability - Langfuse
langfuse>=2.0.0

# Existing dependencies (from pyproject.toml)
requests>=2.32.3
httpx>=0.24.0
